requirements.txt
django>=4.0
django-cors-headers>=3.10.1
django-debug-toolbar>=3.2.4
djangorestframework>=3.13.1
djangorestframework-simplejwt>=5.0.0
django-appconf>=1.0.5
django-simple-history>=3.0.0
django-taggit>=2.0.0
django-filter>=21.1
mysqlclient>=2.1.0
phonenumbers>=8.12.40
pyotp>=2.6.0
requests>=2.27.1
celery>=5.2.3
Pillow>=9.0.0
wheel>=0.37.1
markdown>=3.3.6
redis>=4.1.1